+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### API changes
 
 - Provide `GET /api/users/me/features` endpoint
+- Accessing `GET /api/users/me` with an invalid user id now returns a `unauthorized` error
 
 ## Ara 2018-09-09